@@ -53,13 +53,12 @@ class MediaEntry < MediaResource
    meta_data_for_context = compared_against.meta_data_for_context(context)
    
    new_blank_media_entry = self.new
-   meta_data_for_context.inject([]) do |meta_data, md_bare|
-      meta_data << if other_entries.any? {|me| not me.meta_data.get(md_bare[:meta_key_id]).same_value?(md_bare[:value])}
-        MetaDatum.new(:media_resource => new_blank_media_entry, :meta_key_id => md_bare[:meta_key_id], :value => nil, :keep_original_value => true)
+   meta_data_for_context.map do |md_bare|
+      if other_entries.any? {|me| not me.meta_data.get(md_bare[:meta_key_id]).same_value?(md_bare[:value])}
+        new_blank_media_entry.meta_data.build(:meta_key_id => md_bare[:meta_key_id], :keep_original_value => true)
       else
-        MetaDatum.new(:media_resource => new_blank_media_entry, :meta_key_id => md_bare[:meta_key_id], :value => md_bare[:value])
+        new_blank_media_entry.meta_data.build(:meta_key_id => md_bare[:meta_key_id], :value => md_bare[:value])
       end
-      meta_data
    end
  end
 
