@@ -68,9 +68,9 @@ class MetaContext < ActiveRecord::Base
 
   # TODO dry with MediaSet#abstract  
   def abstract(current_user = nil, min_media_entries = nil)
-    accessible_media_entry_ids = media_entries(current_user).map(&:id)
+    accessible_media_entry_ids = media_entries(current_user).pluck("media_resources.id")
     min_media_entries ||= accessible_media_entry_ids.size.to_f * 50 / 100
-    meta_key_ids = meta_keys.where(:is_dynamic => nil).map(&:id) # TODO get all related meta_key_ids ?? 
+    meta_key_ids = meta_keys.where(:is_dynamic => nil).pluck("meta_keys.id") # TODO get all related meta_key_ids ?? 
 
     h = {} #1005# TODO upgrade to Ruby 1.9 and use ActiveSupport::OrderedHash.new
     mds = MetaDatum.where(:meta_key_id => meta_key_ids, :media_resource_id => accessible_media_entry_ids)
@@ -101,10 +101,10 @@ class MetaContext < ActiveRecord::Base
 
   # TODO dry with MediaSet#used_meta_term_ids  
   def used_meta_term_ids(current_user = nil)
-    meta_key_ids = meta_keys.for_meta_terms.map(&:id)
+    meta_key_ids = meta_keys.for_meta_terms.pluck("meta_keys.id")
 
     mds = if current_user
-      accessible_media_entry_ids = MediaEntry.accessible_by_user(current_user).map(&:id)
+      accessible_media_entry_ids = MediaEntry.accessible_by_user(current_user).pluck(:id)
       MetaDatum.where(:meta_key_id => meta_key_ids, :media_resource_id => accessible_media_entry_ids)
     else
       MetaDatum.where(:meta_key_id => meta_key_ids)
