@@ -14,7 +14,7 @@ class MetaKey < ActiveRecord::Base
   has_many :meta_contexts, :through => :meta_key_definitions
 
 #old#
-#  has_and_belongs_to_many :meta_terms, :class_name => "MetaTerm",  # TODO enforce object_type="MetaTerm" if meta_terms
+#  has_and_belongs_to_many :meta_terms, :class_name => "MetaTerm",  # TODO enforce meta_datum_object_type="MetaDatumMetaTerms" if meta_terms
 #                                       :join_table => :meta_keys_meta_terms,
 #                                       :association_foreign_key => :meta_term_id
   has_many :meta_key_meta_terms
@@ -25,18 +25,14 @@ class MetaKey < ActiveRecord::Base
 
   #old#precedence problem# default_scope order(:label)
   scope :with_meta_data, joins(:meta_data).group(:id)
-  scope :for_meta_terms, where(:object_type => "MetaTerm") 
+  scope :for_meta_terms, where(:meta_datum_object_type => "MetaDatumMetaTerms") 
   
 ########################################################
 
-  before_save do
-    self.object_type = nil if object_type.blank?
-  end
-
   before_update do
-    if object_type_changed?
-      case object_type
-        when "MetaTerm"
+    if meta_datum_object_type_changed?
+      case meta_datum_object_type
+        when "MetaDatumMetaTerms"
           self.is_extensible_list = true
           meta_data.each {|md| md.update_attributes(:value => md.value) }
         # TODO when... else
@@ -64,12 +60,6 @@ class MetaKey < ActiveRecord::Base
   def key_map_for(context)
     d = meta_key_definitions.for_context(context)
     d.key_map if d
-  end
-
-########################################################
-
-  def object_class
-    object_type.constantize
   end
 
 ########################################################
@@ -110,14 +100,14 @@ class MetaKey < ActiveRecord::Base
   end
 
   def self.object_types
-    where("object_type IS NOT NULL").group(:object_type).collect(&:object_type).sort
+    where("meta_datum_object_type IS NOT NULL").group(:meta_datum_object_type).collect(&:meta_datum_object_type).sort
   end
   
 ########################################################
 
   # TODO refactor to association has_many :used_meta_terms, :through ...
   def used_term_ids
-    meta_data.flat_map(&:value).uniq.compact if object_type == "MetaTerm"
+    meta_data.flat_map(&:value).uniq.compact if meta_datum_object_type == "MetaDatumMetaTerms"
   end
 
 end
