@@ -107,7 +107,7 @@ end
 
 Then /^the counter is formatted as "([^"]*)"$/ do |string|
   re = Regexp.new(string.gsub(/[N,M]/,'\d')) 
-  wait_until { find("*", :text => re) }
+  wait_until(15){ find("section", :text => re) }
 end
 
 Given /^the system is set up$/ do
