@@ -5,11 +5,13 @@ describe MetaDataController do
 
 
   before :all do
+    FactoryGirl.create :usage_term 
+    FactoryGirl.create :meta_context_core
     @user = FactoryGirl.create :user
     @other_user = FactoryGirl.create :user
     @media_set= FactoryGirl.create :media_set, user: @user
     @media_set.meta_data.create(:meta_key => MetaKey.find_by_label("title"), :value => Faker::Lorem.words(4).join(' '))
-    @title_meta_datum =  @media_set.meta_data.joins(:meta_key).where("meta_keys.label = ?","title").first
+    @title_meta_datum =  @media_set.meta_data.joins(:meta_key).where(:meta_keys => {:label => "title"}).first
   end
 
   let :valid_session do
@@ -47,13 +49,13 @@ describe MetaDataController do
     it "should contain the correct title meta_datum" do
       json = get_json_as_hash
       json.detect{|e| e["name"] == "title"}.should be
-      json.detect{|e| e["name"] == "title"}["type"].should == "String"
+      json.detect{|e| e["name"] == "title"}["type"].should == "string"
     end
 
     it "should contain the correct owner meta_datum" do
       json = get_json_as_hash
       json.detect{|e| e["name"] == "owner"}.should be
-      json.detect{|e| e["name"] == "owner"}["type"].should == "User"
+      json.detect{|e| e["name"] == "owner"}["type"].should == "users"
     end
   end
 
