@@ -4,15 +4,8 @@ module DataFactory
   extend self
 
   def reset_data 
-    ActiveRecord::Base.transaction do
-      MediaResource.destroy_all
-      Grouppermission.destroy_all
-      Group.destroy_all
-      Userpermission.destroy_all
-      User.destroy_all
-    end
-    MetaHelper.import_initial_metadata
-    Copyright.init(true)
+    DatabaseCleaner.clean_with :truncation
+    DevelopmentHelpers::MetaDataPreset.load_minimal_yaml
   end
 
   def create_permission_migration_dataset
@@ -92,100 +85,6 @@ module FactoryHelper
     bias = ((opts and opts[0]) or 0.5)
     raise "bias must be a real number within [0,1)" if bias < 0.0 or bias >= 1.0
     (rand < bias) ? true : false
-  end
-
-end
-
-FactoryGirl.define do
-
-  ### Media ....
-
-  factory :media_resource_arc , :class => MediaResourceArc do
-  end
-
-  factory :media_file  do 
-    uploaded_data  {
-      f = "#{Rails.root}/features/data/images/berlin_wall_01.jpg"
-      ActionDispatch::Http::UploadedFile.new(:type=> Rack::Mime.mime_type(File.extname(f)),
-                                             :tempfile=> File.new(f, "r"),
-                                             :filename=> File.basename(f))
-    } 
-  end
-
-
-  ### Permissions ...
-
-
-  factory :permission do
-    subject {FactoryHelper.rand_bool ? User.find_random : Group.find_random }
-    media_resource { MediaResource.find_random }
-    after_build do |perm| 
-      permissions = {:view => FactoryHelper.rand_bool, :edit => FactoryHelper.rand_bool, :manage => FactoryHelper.rand_bool, :download => FactoryHelper.rand_bool}
-      perm.set_actions permissions 
-    end
-  end
- 
-
-
-
-  factory :userpermission do
-    view {FactoryHelper.rand_bool 1/4.0}
-    download { view and FactoryHelper.rand_bool}
-    edit {FactoryHelper.rand_bool 1/4.0}
-    manage {edit and FactoryHelper.rand_bool}
-    user {User.find_random || (FactoryGirl.create :user)} 
-
-    media_resource do 
-      MediaResource.find_random || 
-        begin
-          if FactoryHelper.rand_bool 1.0/3
-            FactoryGirl.create :media_set
-          else
-            FactoryGirl.create :media_entry
-          end
-        end
-    end
-
-  end
-
-
-  factory :grouppermission do
-
-    view {FactoryHelper.rand_bool 1/4.0}
-    download { view and FactoryHelper.rand_bool}
-    edit {FactoryHelper.rand_bool 1/4.0}
-    manage {edit and FactoryHelper.rand_bool}
-
-    group {Group.find_random || (FactoryGirl.create :group)}
-
-    media_resource do 
-      if FactoryHelper.rand_bool 1.0/3
-        FactoryGirl.create :media_set
-      else
-        FactoryGirl.create :media_entry
-      end
-    end
-
-  end
-
-
-
-  ### Groups, Users, ....
-
-  factory :group do
-    name {Faker::Name.last_name}
-  end
-
-  factory :person do
-    lastname {Faker::Name.last_name}
-    firstname {Faker::Name.first_name}
-  end
-
-  factory :user do |n| 
-    person {FactoryGirl.create :person}
-    email {Faker::Internet.email}
-    login {Faker::Internet.user_name + "_#{UUIDTools::UUID.random_create.to_s.first 5}"}
-    usage_terms_accepted_at {Time.now}
   end
 
 end
