@@ -4,5 +4,5 @@ FactoryGirl.define do
   factory :media_resource do
     user {User.find_random || (FactoryGirl.create :user)}
   end
-
+  
 end
