@@ -1,5 +1,7 @@
 FactoryGirl.define do
   factory :meta_key do
-    label {Faker::Lorem.words.join("_")}
+    label { Faker::Lorem.words.join("_") }
+    meta_datum_object_type { "MetaDatumString" }
+    is_dynamic { nil }
   end
 end
